cleanup: local data and functions can be static (tMenu)
Most of the changes in this commit are making code that is only used
in this file (compilation unit) static. Careful reading the changes show
a little more. Like, removing commented code (as any code change can be
found the repository, so nothing is really lost), or removing
obvious looking useless code (callerID).

Signed-off-by: Jan Mulder <jlmulder@xs4all.nl>