Fix accidentally pushed local code
Did I say that Mercurial is a pain for long time git users :-)

For the record: created a commit with 1 file, saw a spelling error
so, did a hg commit --amend ... which sneaked in a 2nd file. I did not
notice, created a pull request, and now this stupid commit undoing things
is the only way as all is in public phase now.

Signed-off-by: Jan Mulder <jlmulder@xs4all.nl>