cleanup: another cleanup session (data_exchange_main.c)
Cleanup all kinds of stuff in one of the core functions. Made local
data static, with 1 already existing exception (the dataIn data). For
this, a trivial change is made in t7.c. In addition, not used code, and
commented out code that seems highly obsolete is removed. Another tiny
step in code that is better to maintain in the future.

Signed-off-by: Jan Mulder <jlmulder@xs4all.nl>