Bugfix: handle rounding of TTS and future TTS consistently
There was a bug with futureTTS with an extended bottom time of 0
minutes (so same TTS and the main schedule) being not in sync with
each other. This was simply caused by inconsistent rounding of
seconds to minute conversions. And when fixing this, I noticed
also inconsistent handling of huge 1000+ minutes TTS values. At
some places, the presentation changed to hours notation, and other
places not. I doubt that any OSTC4 user ever made a dive that had
more than 1000 minutes of deco, but fixed this presentation as
well, simply because of consistency.

Signed-off-by: Jan Mulder <jlmulder@xs4all.nl>