feature: replace Relative GF by saturation, computational only
Replace Relative GF by saturation. The saturation code is derived from
the hwOS repo. This commit only contains the Buhlmann computational part. In
the simulator, the numbers does look sane. The initial ascent from the
bottom phase stops at the deepest deco stop a little over GFlow, which is
correct in my view, as we still have some time to spend on this deepest
stop, and that stop ends very close to GFlow (obviously, a deepest stop
is typically short, as in 1 or 2 minutes). The deco finally ends at the
surface with a saturation value of GFhigh.

Signed-off-by: Jan Mulder <jlmulder@xs4all.nl>