bugfix, simulator: fix the 1 sec difference between stopwatch and divetime
This is a rather subtle bugfix. counterSecondsShallowDepth is used to decide
to increase dive_time_seconds_without_surface_time. But just after using
it, logic is present to reset counterSecondsShallowDepth. So, basically,
we are looking to an old state.

Simply increase dive_time_seconds_without_surface_time *after* we have reset
counterSecondsShallowDepth, so that we are looking to a current state instead
of an old one.

Notice that this is simulator only.

Signed-off-by: Jan Mulder <jlmulder@xs4all.nl>